aCommit

+++ pkg/sd/item.go
@@ -3,5 +3,5 @@ package sd
 // Item represents a single attribute and its value in a row
 type Item struct {
 	Attr  string
-	Value float64
+	Value int
 }
